app/requirements.txt
boto3==1.34.132   
botocore==1.34.132   
certifi==2024.6.2   
charset-normalizer==3.3.2      
idna==3.7        
jmespath==1.0.1      
pip==24.1       
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gitlab==4.9.0
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.2
setuptools==65.5.0
six==1.16.0
slack_bolt==1.19.0
slack_sdk==3.30.0
tabulate==0.9.0
urllib3==2.2.2